Cache reduction to ~99GiB via the -a/--alternate switch (#190)
* Added cache reduction to ~99GiB via the -a/--alternate switch

* Fixing regression on getting disk block size on windows.

* Fixed f1 time not accumulating last I/O wait to the correct accumulator.

* Fixes other minor issues

+++ src/plotdisk/k32/DiskPlotBounded.cpp
@@ -42,12 +42,12 @@ K32BoundedPhase1::K32BoundedPhase1( DiskPlotContext& context )
 
     // Temp2
     {
-        FileSetOptions opts = FileSetOptions::Interleaved;
+        FileSetInitData data = {};
+        FileSetOptions  opts = context.cfg->alternateBuckets ? FileSetOptions::Alternating : FileSetOptions::Interleaved;
         
         if( !context.cfg->noTmp2DirectIO )
             opts |= FileSetOptions::DirectIO;
 
-        FileSetInitData data = {};
         opts |= FileSetOptions::UseTemp2;
 
         size_t metaCacheSize = 0;
@@ -55,8 +55,7 @@ K32BoundedPhase1::K32BoundedPhase1( DiskPlotContext& context )
         if( context.cache )
         {
             // In fully interleaved mode (bigger writes chunks), we need 192GiB for k=32
-            // In alternating mode, we need 96 GiB.
-            // #TODO: Support alternating mode
+            // In alternating mode, we need 96 GiB (around 99GiB if we account for disk block-alignment requirements).
             
             opts |= FileSetOptions::Cachable;
             data.cache = context.cache;
@@ -64,9 +63,9 @@ K32BoundedPhase1::K32BoundedPhase1( DiskPlotContext& context )
             // Proportion out the size required per file:
             //  A single y or index file requires at maximum 16GiB
             //  A asingle meta file at its maximum will require 64GiB
-            // Divide the whole cache into 12 equal parts, where each meta file represent 4 parts.
+            // Divide the whole cache into 12 (6 for alternating mode) equal parts, where each meta file represent 4 parts.
             // Ex: 192 / 12 = 16.  This gives us 4 files of 16GiB and 2 files of 64GiB
-            size_t singleFileCacheSize = context.cacheSize / 12;
+            size_t singleFileCacheSize = context.cacheSize / ( context.cfg->alternateBuckets ? 6 : 12 );
 
             // Align to block size
             singleFileCacheSize = numBuckets * RoundUpToNextBoundaryT( singleFileCacheSize / numBuckets - context.tmp2BlockSize, context.tmp2BlockSize );
@@ -83,14 +82,38 @@ K32BoundedPhase1::K32BoundedPhase1( DiskPlotContext& context )
             data.cache = (byte*)data.cache + data.cacheSize;
         };
 
-        InitCachableFileSet( FileId::FX0   , "y0"    , numBuckets, opts, data );
-        InitCachableFileSet( FileId::FX1   , "y1"    , numBuckets, opts, data );
-        InitCachableFileSet( FileId::INDEX0, "index0", numBuckets, opts, data );
-        InitCachableFileSet( FileId::INDEX1, "index1", numBuckets, opts, data );
+        if( _context.cfg->alternateBuckets )
+        {
+            const uint64 blockSize       = context.tmp2BlockSize;
+            const uint64 tableEntries    = 1ull << 32;
+            const uint64 bucketEntries   = tableEntries / numBuckets;
+            const uint64 sliceEntries    = bucketEntries / numBuckets;
 
-        data.cacheSize = metaCacheSize;
-        InitCachableFileSet( FileId::META0, "meta0", numBuckets, opts, data );
-        InitCachableFileSet( FileId::META1, "meta1", numBuckets, opts, data );
+            const uint64 ysPerBlock      = blockSize / sizeof( uint32 );
+            const uint64 metasPerBlock   = blockSize / (sizeof( uint32 ) * 4);
+
+            const uint64 sliceSizeY    = RoundUpToNextBoundaryT( (uint64)(sliceEntries * BB_DP_ENTRY_SLICE_MULTIPLIER), ysPerBlock    ) * sizeof( uint32 );
+            const uint64 sliceSizeMeta = RoundUpToNextBoundaryT( (uint64)(sliceEntries * BB_DP_ENTRY_SLICE_MULTIPLIER), metasPerBlock ) * sizeof( uint32 ) * 4;
+            
+            data.maxSliceSize = sliceSizeY;
+            InitCachableFileSet( FileId::FX0   , "y0"    , numBuckets, opts, data );
+            InitCachableFileSet( FileId::INDEX0, "index0", numBuckets, opts, data );
+
+            data.maxSliceSize = sliceSizeMeta;
+            data.cacheSize        = metaCacheSize;
+            InitCachableFileSet( FileId::META0, "meta0", numBuckets, opts, data );
+        }
+        else
+        {
+            InitCachableFileSet( FileId::FX0   , "y0"    , numBuckets, opts, data );
+            InitCachableFileSet( FileId::FX1   , "y1"    , numBuckets, opts, data );
+            InitCachableFileSet( FileId::INDEX0, "index0", numBuckets, opts, data );
+            InitCachableFileSet( FileId::INDEX1, "index1", numBuckets, opts, data );
+
+            data.cacheSize = metaCacheSize;
+            InitCachableFileSet( FileId::META0, "meta0", numBuckets, opts, data );
+            InitCachableFileSet( FileId::META1, "meta1", numBuckets, opts, data );
+        }
     }
 }
 
@@ -126,8 +149,6 @@ size_t K32BoundedPhase1::GetRequiredSize( const uint32 numBuckets, const size_t 
 //-----------------------------------------------------------
 void K32BoundedPhase1::Run()
 {
-    Log::Line( "Table 1: F1 generation" );
-
     switch( _context.numBuckets )
     {
         case 64 : RunWithBuckets<64 >(); break;
@@ -209,16 +230,22 @@ void K32BoundedPhase1::RunWithBuckets()
         #if !BB_DP_P1_KEEP_FILES
             if( table == TableId::Table6 )
             {
-                // #TODO: Support alternating mode (only 1 metadata file)
-                _ioQueue.DeleteBucket( FileId::META0 );
-                _ioQueue.CommitCommands();
+                if( !_context.cfg->alternateBuckets )
+                {
+                    _ioQueue.DeleteBucket( FileId::META0 );
+                    _ioQueue.CommitCommands();
+                }
             }
             else if( table == TableId::Table7 )
             {
-                // #TODO: Support alternating mode (only 1 metadata file)
-                _ioQueue.DeleteBucket( FileId::FX1    );
-                _ioQueue.DeleteBucket( FileId::INDEX1 );
-                _ioQueue.DeleteBucket( FileId::META1  );
+                if( _context.cfg->alternateBuckets )
+                    _ioQueue.DeleteBucket( FileId::META0 );
+                else
+                {
+                    _ioQueue.DeleteBucket( FileId::FX1    );
+                    _ioQueue.DeleteBucket( FileId::INDEX1 );
+                    _ioQueue.DeleteBucket( FileId::META1  );
+                }
                 _ioQueue.CommitCommands();
             }
         #endif
@@ -241,11 +268,9 @@ void K32BoundedPhase1::RunWithBuckets()
     }
 
      #if !BB_DP_P1_KEEP_FILES
-        // #TODO: Support alternating mode (only 1 metadata file)
         _ioQueue.DeleteBucket( FileId::FX0    );
         _ioQueue.DeleteBucket( FileId::INDEX0 );
         _ioQueue.CommitCommands();
-
         // # TODO: Wait for deletion?
     #endif
 }
@@ -256,14 +281,15 @@ void K32BoundedPhase1::RunF1()
 {
     _context.ioWaitTime = Duration::zero();
 
+    Log::Line( "Table 1: F1 generation" );
     Log::Line( "Generating f1..." );
-    auto timer = TimerBegin();
 
+    const auto timer = TimerBegin();
     StackAllocator allocator( _context.heapBuffer, _context.heapSize );
     K32BoundedF1<_numBuckets> f1( _context, allocator );
     f1.Run();
+    const double elapsed = TimerEnd( timer );
 
-    double elapsed = TimerEnd( timer );
     Log::Line( "Finished f1 generation in %.2lf seconds. ", elapsed );
     Log::Line( "Table 1 I/O wait time: %.2lf seconds.", _context.p1TableWaitTime[(int)TableId::Table1] );
     
@@ -315,10 +341,11 @@ void K32BoundedPhase1::RunFx()
 
     #if _DEBUG
         BB_DP_DBG_WriteTableCounts( _context );
-
+         // #TODO: Update this for alternating mode
         _ioQueue.DebugWriteSliceSizes( table, (uint)table %2 == 0 ? FileId::FX0    : FileId::FX1    );
         _ioQueue.DebugWriteSliceSizes( table, (uint)table %2 == 0 ? FileId::INDEX0 : FileId::INDEX1 );
         _ioQueue.DebugWriteSliceSizes( table, (uint)table %2 == 0 ? FileId::META0  : FileId::META1  );
         _ioQueue.CommitCommands();
     #endif
 }
+
